Implement Grove sound sensor
A fun test is to open the Python interpreter, instantiate a
GroveDevices.Grove_Sound_Sensor to s, and run
`while s.soundIntensity() > 2000: pass`
The cursor will stay in place and the prompt will not appear until there
is a sound. You can blow air towards the sensor from about 8 or 10
inches to stop the loop.