Add optional type parameter to GroveSensor
Makes the initializer of Grove_Digital_Sensor and Grove_Analog_Sensor
cleaner in my opinion. Parameter is optional because before it was not
required (backwards-compatibility). verifyPort was not merged in to the
initializer for GroveSensor because the type must be set before it can
be verified, and that is not required at init time.
Also, the mode may change later!