Make RCServo more like GroveSensor
Instead of being instantiated in the PiStorms_GRX constructor and being
properties of that class, users will create RCServo instances. This is
required because the encoder port will need to be (optionally)
specified. This could have been done with a setEncoderPort as with
setNeutralPoint (set in __init__, can be changed later), but it seems to
make more sense this way.