Retain old data when displaying latest subset
Only updated in 01-GraphGO, 03-MultipleAxis still discards old data.
When you are just appending data and call plot again, it updates the
line. However, if the data you're plotting doesn't start with the same
data that's there, it draws a new line instead of updating the old one.
One way to go about this is to call set_ydata on the line, or remove the
old line and plot a new one. For this simple example the latter is more
clear.
02-SingleSensor removed because it doesn't add anything.