Clean NXTSERVO.setNeutral
Downloading the NXT neutral utility from our website, setNeutral is
actually implemented the same way! I was going to test this with a
continuous rotation servo, but setNeutral only accepts a single argument
so I can't tell how to use or test it. Also I got a NameError: name
'time' is not defined, but time.sleep is called in 13 other places in
this file! I'm shocked the time module was not imported, would
attempting to use any of these other classes have thrown a NameError
too?