Add polar graphic example to data visualization
I wanted to have this use the threaded data collection and fast redrawing, but I ran in to some issues. Particularly I wanted a smooth line, but the function won't interpolate through 0, so instead it makes long arcs that look bad. In the end I went with a more basic example.