Correct type checking for GroveAnalogPort and remove setType
An instance of GroveDigitalPort or GroveAnalogPort will be created when
a class from GroveDevices is initialized. The port type that sensor will
use should not change. However, you can instantiate another class with
that port as an argument and it will change the type if needed. Even
more, the method is accessible through portInstace.comm.setType