Use sensors included in Grove Sensor Kit in Blockly example
I wanted to use sensors that sound cool for the example. The EV3 has a
touch, sound, and light sensor, but the luminance, loudness and UV
sensors are unique to the GRX.
We can't expect everyone to have those sensors though, so now the
example will use those included in the Grove Sensor Kit product.