Force intermediate steps in drawButton to not refresh display
The drawButton performs 4 or 5 drawing operations. It will always do
three steps for drawing the left, center, and right of the button, and
one step to draw teh text on the button. It might also draw an icon on
the image before this last step. Previously the intermediate 3 or 4
steps passed-through the display argument. However, there is no reason
to update the display when only the left, left and center, etc. of the
button has been drawn. A valid option would be to refresh the screen
when the entire button is drawn, which remains.