Use .getKeyPressCount to exit data viz examples with a tap
Previously, when using isKeyPressed, you would have to tap and hold the GO button for a moment to quit the program. This was to make sure you are pressing it when isKeyPressed is checked between iterations. By waiting for getKeyPressCount to return at least one, tapping between iterations is fine (the PiStorms keeps track of how many times the GO button has been pressed). Instantiating a PiStorms instance (as at the top of each program) resets the count.